requirements.txt
cffi==1.15.1
cryptography==38.0.4
idna==3.4
pycparser==2.21
pyOpenSSL==22.1.0
